requirements.txt
pymodbus==3.5.1
pyserial==3.5
streamlit==1.35.0
paho-mqtt==1.6.1
python-telegram-bot==20.8
requests==2.31.0
websockets==12.0
psutil==5.9.8
crcmod==1.7
plotly==5.18.0
pandas==2.1.4
